% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{calcActivationDifference}
\alias{calcActivationDifference}
\title{Calculate activation difference between classes.}
\usage{
calcActivationDifference(data)
}
\arguments{
\item{data}{Tibble containing columns labelled Filter, Y, Activation, Effect, and Information.}
}
\value{
Tibble with mean activation difference calculated.
}
\description{
Per first-layer convolutional filter, calculate the mean activation difference between class labels.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{activation}
\keyword{difference}
\keyword{filter}
