% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_offsets.R
\name{calculate_offsets}
\alias{calculate_offsets}
\title{Calculate offset}
\usage{
calculate_offsets(filter.motifs, motif.pos = NULL, motif.maxlen,
  onehot_array, batchsize = 256)
}
\arguments{
\item{filter.motifs}{4-D array of ICMs.}

\item{motif.pos}{Vector of indices to extract motifs. If \code{NULL} all motifs will be selected.}

\item{motif.maxlen}{Maximum motif length, found via \code{\link{find_max_length}}.}

\item{onehot_array}{One-hot encoded training data to sample a batch from.}

\item{batchsize}{Size of batch to sample.}
}
\value{
Vector of offsets of length equal to the fourth dimension of the input array of ICMs.
}
\description{
Calculate the offset (bias) as the mean activation over a sampling of the data.
}
\seealso{
\code{\link{find_max_length}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{bias}
\keyword{offset}
