% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_definitions.R
\name{combo_constraint}
\alias{combo_constraint}
\title{Combination constraint to set non-negative and fix non-de novo weights.}
\usage{
combo_constraint(w)
}
\arguments{
\item{w}{Weights.}
}
\value{
Constraint.
}
\description{
Constrains weights to be non-negative and fixes the weights for the fixed motifs.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{combo}
\keyword{constraint}
