% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_IGM.R
\docType{data}
\name{combo_IGM_Motifs}
\alias{combo_IGM_Motifs}
\title{Model callback function wrapper for learning IGMs along with using fixed motifs..}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
combo_IGM_Motifs
}
\description{
Callback for the model to output plots after a user-specified number of epochs. Model must have two input layers and utilize IGMs as filters.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{callback}
\keyword{model}
