% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_PWM.R
\docType{data}
\name{combo_PWM_Motifs}
\alias{combo_PWM_Motifs}
\title{Model callback function wrapper for learning PWMs along with using fixed motifs.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
combo_PWM_Motifs
}
\description{
Callback for the model to output plots after a user-specified number of epochs. Model must have two input layers and utilize PWMs as filters.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{callback}
\keyword{model}
