% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_definitions.R
\docType{data}
\name{deNovo_Motifs}
\alias{deNovo_Motifs}
\title{Model callback function wrapper.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
deNovo_Motifs
}
\description{
Callback for the model to output plots after a user-specified number of epochs. Model must only have a single input layer.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{callback}
\keyword{model}
