% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EmptyBackground}
\alias{EmptyBackground}
\title{Nucleotide sequences sampled from random background.}
\format{A data frame with 1000 rows and 3 variables:
\describe{
  \item{seqName}{Sequence identifier}
  \item{sequence}{Nucleotide sequences}
  \item{embeddings}{Locations and descriptions of embedded motifs.}
}}
\source{
\url{https://github.com/kundajelab/simdna}
}
\usage{
EmptyBackground
}
\description{
A dataset containing nucleotide sequences of length 200 basepairs composed of nucleotides A, C, G, T.
Each sequence is sampled from random genome background with 60% GC-content.
}
\keyword{datasets}
