% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizer_defintions.R
\name{filter_regularizer}
\alias{filter_regularizer}
\title{Sparse group lasso penalty applied to convolutional filters.}
\usage{
filter_regularizer(weight_matrix)
}
\arguments{
\item{weight_matrix}{Filter weights.}
}
\value{
Sparse group lasso penalty.
}
\description{
L1 regularization penalty to constrain individual filter weights plus the L1/L2 regularization to collectively regularize the entire convolutional filter as a group.
}
\seealso{
\code{\link{position_regularizer}} \code{\link{total_regularizer}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{filter}
\keyword{group}
\keyword{regularizer}
\keyword{sparse}
