% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_annotated_motifs.R
\name{find_max_length}
\alias{find_max_length}
\title{Find maximum motif length.}
\usage{
find_max_length(motif.raw, motif.pos)
}
\arguments{
\item{motif.raw}{List of annotated motifs as PFMs.}

\item{motif.pos}{Vector of indices to extract motifs. If \code{NULL} all motifs will be selected.}
}
\value{
Length of the longest motif.
}
\description{
Loop through all motifs and return the length of the longest motif.
}
\seealso{
\code{\link{pfm_to_icm}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{content}
\keyword{icm}
\keyword{information}
\keyword{motif}
\keyword{pfm}
