% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_definitions.R
\name{fixedMotif_constraint}
\alias{fixedMotif_constraint}
\title{Constrains weights to fixed motifs.}
\usage{
fixedMotif_constraint(w)
}
\arguments{
\item{w}{Weights.}
}
\value{
Constraint.
}
\description{
Constrains weights to to the fixed motifs. These may be the annotated database motifs or perhaps de novo motifs.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{constraint}
\keyword{fixed}
\keyword{motif}
