% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_motifs.R
\name{format_activation_motifs}
\alias{format_activation_motifs}
\title{Find maximally-activating sequences for motif plotting.}
\usage{
format_activation_motifs(activations, test_seqs, filter_len, method, threshold)
}
\arguments{
\item{activations}{4-D array of activation values per filter.}

\item{test_seqs}{Data to extract nucleotide sequences from. Must be a character array.}

\item{filter_len}{Length of the sequence motif to extract; the length of the convolutional filters.}

\item{method}{Either \code{'fixed'} (extract the sequences with activation greater than \code{threshold*max_activation}),
\code{'max-pooled'} (extract the sequences which gave rise to the max-pooled value), or
\code{'quantile'} (extract the sequences with activation greater than \code{quantile(activations,probs=threshold)}.}

\item{threshold}{Threshold above which to extract sequences.}
}
\value{
List with named motifs.
}
\description{
Find the sequeneces which maximally activate (maximially cross-correlated with) the filters and format to a list with named rows for plotting with \code{\link{plot_motifs}}.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{filter}
\keyword{format}
\keyword{motif}
