% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{format_activations}
\alias{format_activations}
\title{Format filter activations for plotting.}
\usage{
format_activations(model, data, y, input_layer = 1, output_layer,
  buffer = 0, motif.names)
}
\arguments{
\item{model}{A \link{\code{Keras}} model.}

\item{data}{Data array returned from \link{\code{one_hot}} to obtain activations for.}

\item{y}{Vector of class labels.}

\item{input_layer}{Integer value denoting the input data. If NULL, assumes a single input layer.}

\item{output_layer}{Integer denoting output layer or string containing the layer name.}

\item{buffer}{Where to start extracting the beta coefficients from.}

\item{motif.names}{Array containing names for the filters.}
}
\value{
Tibble of formatted activations.
}
\description{
Format the first-layer filter activations for plotting with \link{\code{plot_filter_activations_byClass}}.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{activations}
\keyword{filter}
\keyword{format}
