% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_motifs.R
\name{format_filter_motifs}
\alias{format_filter_motifs}
\title{Format filters for motif plotting.}
\usage{
format_filter_motifs(conv_filters, type = "IGM", bg = c(0.25, 0.25,
  0.25, 0.25))
}
\arguments{
\item{conv_filters}{4-dimensional array containing filter weights.}

\item{type}{String declaring motif representation type (either `IGM` or `PWM`).}

\item{bg}{Vector of background nucleotide probabilities (A, C, G, T)}
}
\value{
List with named motifs.
}
\description{
Convert the information content motifs to a list with named rows for plotting with \code{\link{plot_motifs}}.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{filter}
\keyword{format}
\keyword{motif}
