% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{get_activations}
\alias{get_activations}
\title{Calculate activations.}
\usage{
get_activations(model, input_layer = NULL, output_layer, data)
}
\arguments{
\item{model}{A \code{Keras} model.}

\item{input_layer}{Integer value denoting the input data. If NULL, assumes a single input layer.}

\item{output_layer}{Integer denoting output layer or string containing the layer name.}

\item{data}{Data array returned from \code{\link{one_hot}} to obtain activations for.}
}
\value{
Array of activations.
}
\description{
Calculate first-layer convolutional filter activations.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{activations}
\keyword{filter}
