% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_definitions.R
\name{info_constraint}
\alias{info_constraint}
\title{Constrains filter weights within information bounds.}
\usage{
info_constraint(w)
}
\arguments{
\item{w}{Filter weights.}
}
\value{
Constraint.
}
\description{
Constrains the individual filters weights to sum column-wise less than or equal to 2.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{constraint}
\keyword{information}
