% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{IUPAC2Matrix}
\alias{IUPAC2Matrix}
\title{Convert dna strings to PFM}
\usage{
IUPAC2Matrix(x)
}
\arguments{
\item{x}{nucleotide strings.}
}
\value{
A matrix.
}
\description{
Converts nucleotide strings to position frequency matrices. Function available in the TFBSTools package (\code{https://github.com/ge11232002/TFBSTools}).
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{biostrings}
\keyword{iupac}
\keyword{tfbstools}
