% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_denovoMotif.R
\name{layer_deNovo}
\alias{layer_deNovo}
\title{Custom 2-D convolution wrapper with to learn ICMs.}
\usage{
layer_deNovo(object, filters, filter_len, lambda_pos = 0, lambda_filter = 0,
  lambda_l1 = 0, lambda_offset = 0, lambda_Z_offset = 0, strides = c(1L,
  1L), padding = "valid", activation = "sigmoid", use_bias = TRUE,
  kernel_initializer = initializer_random_uniform(0, 0.5),
  bias_initializer = "zeros", kernel_regularizer = total_regularizer,
  activity_regularizer = Z_regularizer, kernel_constraint = info_constraint,
  bias_constraint = negative_constraint, input_shape = NULL,
  name = "deNovo_conv", trainable = NULL)
}
\arguments{
\item{object}{Keras model object.}

\item{filters}{Number of convolutional filters.}

\item{filter_len}{Length of the motifs.}

\item{lambda_pos}{Position-wise L1 sparsity penalty.}

\item{lambda_filter}{Sparse group lasso penalty.}

\item{lambda_l1}{Sparse L1 penalty applied to filter weights.}

\item{lambda_offset}{L1 penalty applied the offsets (bias) terms}

\item{lambda_Z_offset}{L1 penalty for activations centered around 0.5.}
}
\value{
A Keras model with the added layer.
}
\description{
Custom 2-D convolution wrapper with regularization and constraints to learn de novo motifs in the form of ICMs.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{conv}
\keyword{convolution}
\keyword{denovo}
\keyword{layer}
