% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_fixedMotif.R
\name{layer_fixedMotif}
\alias{layer_fixedMotif}
\title{Custom 2-D convolution wrapper with fixed ICMs}
\usage{
layer_fixedMotif(object, filters, motif_maxlen, fixed.motifs, fixed.offsets,
  input_shape = NULL, strides = c(1L, 1L), padding = "valid",
  activation = "sigmoid", use_bias = TRUE, name = "fixedMotif_conv",
  trainable = FALSE)
}
\arguments{
\item{object}{Keras model object.}

\item{filters}{Number of convolutional filters.}

\item{motif_maxlen}{Maximum length of the motifs.}

\item{fixed.motifs}{List of motifs to be used as filters.}

\item{fixed.offsets}{Vector of offset values.}
}
\value{
A Keras model with the added layer.
}
\description{
Custom 2-D convolution wrapper for fixed motifs, such as those coming from a database.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{conv}
\keyword{convolution}
\keyword{fixedMotif}
\keyword{layer}
