% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_topN_motif_hits.R
\name{layer_topN_max_pooling}
\alias{layer_topN_max_pooling}
\title{Top max-pooled hits}
\usage{
layer_topN_max_pooling(object, n_max, name = NULL)
}
\arguments{
\item{object}{Keras model.}

\item{n_max}{Number of top hits to return.}

\item{name}{Layer name (optional).}
}
\value{
Top max-pooled hits layer.
}
\description{
Calculate the top N motif hits max-pooled over the entire sequence.
}
\seealso{
\code{\link{topN_max_pooling}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{hits}
\keyword{max}
\keyword{motif}
\keyword{pooling}
