% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_epoch_dir}
\alias{make_epoch_dir}
\title{Create output directory for current training epoch.}
\usage{
make_epoch_dir(log_dir, epoch, wts)
}
\arguments{
\item{log_dir}{Initial log directory.}

\item{epoch}{Current training epoch.}

\item{wts}{Current weights to output.}
}
\value{
NULL. Epoch directory will be initialized.
}
\description{
Creates the epoch output directory.
}
\seealso{
\code{\link{setup_log_dir}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{directory}
\keyword{epoch}
\keyword{log}
