% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hocomoco_utils.R
\name{my_readHOCOMOCOMatrix}
\alias{my_readHOCOMOCOMatrix}
\title{Read HOCOMOCO file in JASPAR-format.}
\usage{
my_readHOCOMOCOMatrix(fn, type = c("individual", "all"))
}
\arguments{
\item{fn}{Filename containing JASPAR-formatted HOCOMOCO PFMs}
}
\value{
List of PFM motifs.
}
\description{
Reads in HOCOMOCO motifs. Slight modification of the original function available in the TFBSTools package (\code{https://github.com/ge11232002/TFBSTools}).
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{annotated}
\keyword{hocomoco}
\keyword{motif}
\keyword{read}
