% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspar_utils.R
\name{my_readJASPARMatrix}
\alias{my_readJASPARMatrix}
\title{Read JASPAR file-format.}
\usage{
my_readJASPARMatrix(fn, type = c("individual", "all"))
}
\arguments{
\item{fn}{Filename containing JASPAR-formatted PFMs}
}
\value{
List of PFM motifs.
}
\description{
Reads in JASPAR format motifs. Slight modification of the original function available in the TFBSTools package (\code{https://github.com/ge11232002/TFBSTools}).
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{annotated}
\keyword{jaspar}
\keyword{motif}
\keyword{read}
