% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MYC_CTCF_IRF}
\alias{MYC_CTCF_IRF}
\title{Nucleotide sequences sampled from known motifs.}
\format{A data frame with 1000 rows and 3 variables:
\describe{
  \item{seqName}{Sequence identifier}
  \item{sequence}{Nucleotide sequences}
  \item{embeddings}{Locations and descriptions of embedded motifs.}
}}
\source{
\url{https://github.com/kundajelab/simdna}
}
\usage{
MYC_CTCF_IRF
}
\description{
A dataset containing nucleotide sequences of length 200 basepairs composed of nucleotides A, C, G, T.
Each sequence contains between 0-3 ocurrences of the MYC, CTCF, or IRF motifs.
}
\keyword{datasets}
