% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_definitions.R
\name{negative_constraint}
\alias{negative_constraint}
\title{Constrains weights to be negative.}
\usage{
negative_constraint(w)
}
\arguments{
\item{w}{Weights.}
}
\value{
Constraint.
}
\description{
Constrains weights to be negative.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{constraint}
\keyword{negative}
