% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\name{one_hot}
\alias{one_hot}
\title{One-hot encoding of vector of DNA nucleotide strings}
\usage{
one_hot(data, zeros_len)
}
\arguments{
\item{data}{Character strings composed of nucleotides A, C, G, T.}

\item{zeros_len}{Integer value denoting number of zeros to pad vector with. Commonly the convolutional filter length - 1.}
}
\value{
Zero-padded array of one-hot encodings.
}
\description{
Inputs vector of strings composed of nucleotides and outputs a zero-padded array of encodings.
}
\examples{
one_hot(rep(paste(sample(c('A','C','G','T'),100,replace=TRUE),collapse=''),50),10)

}
\seealso{
\code{\link{one_hot_vector}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{one-hot}
