% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\name{one_hot_generator}
\alias{one_hot_generator}
\title{One-hot encoding fit generator}
\usage{
one_hot_generator(X_data, Y_data, batch_size, zeros_len, max_len)
}
\arguments{
\item{X_data}{Character strings composed of nucleotides A, C, G, T.}

\item{Y_data}{Vector of labels.}

\item{batch_size}{Integer value denoting batch size.}

\item{zeros_len}{Integer value denoting number of zeros to pad vector with. Commonly the convolutional filter length - 1.}

\item{max_len}{Maximum sequence length. Calculated per batch if NULL.}
}
\value{
Zero-padded array of one-hot encodings.
}
\description{
Streaming generator to one-hot encode samples during training or predicting.
}
\seealso{
\code{\link{one_hot_vector}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{generator}
\keyword{one-hot}
