% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_annotated_motifs.R
\name{pfm_to_icm}
\alias{pfm_to_icm}
\title{Convert PFM to ICM.}
\usage{
pfm_to_icm(motif.raw, motif.pos, mmax)
}
\arguments{
\item{motif.raw}{List of annotated motifs as PFMs.}

\item{motif.pos}{Vector of indices to extract motifs. If \code{NULL} all motifs will be selected.}

\item{mmax}{Maximum motif length, found via \code{\link{find_max_length}}.}
}
\value{
4-D array containing ICMs of the extracted motifs.
}
\description{
Convert position frequency matrix into an information content matrix to be consistent with the de novo filters.
}
\seealso{
\code{\link{find_max_length}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{content}
\keyword{icm}
\keyword{information}
\keyword{motif}
\keyword{pfm}
