% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_annotated_motifs.R
\name{pfm_to_igm}
\alias{pfm_to_igm}
\title{Convert PFM to IGM.}
\usage{
pfm_to_igm(motif.raw, motif.pos = NULL, mmax, bg = c(0.25, 0.25, 0.25,
  0.25))
}
\arguments{
\item{motif.raw}{List of annotated motifs as PFMs.}

\item{motif.pos}{Vector of indices to extract motifs. If \code{NULL} all motifs will be selected.}

\item{mmax}{Maximum motif length, found via \code{\link{find_max_length}}.}

\item{bg}{Vector of background nucleotide probabilities (A, C, G, T)}
}
\value{
4-D array containing IGMs of the extracted motifs.
}
\description{
Convert position frequency matrix into an information content matrix to be consistent with the de novo filters.
}
\seealso{
\code{\link{find_max_length}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{IGM}
\keyword{content}
\keyword{information}
\keyword{motif}
\keyword{pfm}
