% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspar_utils.R
\name{PFMatrix}
\alias{PFMatrix}
\title{Create PFMatrix object.}
\usage{
PFMatrix(ID = "Unknown", name = "Unknown", matrixClass = "Unknown",
  strand = "+", bg = c(A = 0.25, C = 0.25, G = 0.25, T = 0.25),
  tags = list(), profileMatrix = matrix())
}
\arguments{
\item{ID}{Identifier}

\item{name}{Motif name}

\item{matrixClass}{Type}

\item{strand}{Strand +/-}

\item{bg}{Background probabilities}

\item{tags}{Tags}

\item{profileMatrix}{Current profileMatrix to update}
}
\value{
ProfileMatrix PFMatrix.
}
\description{
Reads in a JASPAR-formatted motif and outputs as a PFMatrix class. Slight modification of the original function available in the TFBSTools package (\code{https://github.com/ge11232002/TFBSTools}).
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{matrix}
\keyword{pfmatrix}
