% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_definitions.R
\name{plot_filter_activations_byClass}
\alias{plot_filter_activations_byClass}
\title{Plot first-layer filter activations by class.}
\usage{
plot_filter_activations_byClass(data, show_beta = F, show_offset = F,
  fl = NULL)
}
\arguments{
\item{data}{Tibble containing effect size information, filter name, mean activation difference, and information.}

\item{show_beta}{Whether to include the estimated effect size estimates as well.}

\item{show_offset}{Whether to include the estimated filter offsets as well.}

\item{fl}{Filename to save plots to. If not provided, a plot object will be returned.}
}
\value{
A ggplot object.
}
\description{
Plot the first-layer convolutional filter activations by class label and filter.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{activations}
\keyword{filter}
\keyword{plot}
