% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_definitions.R
\name{plot_motifs}
\alias{plot_motifs}
\title{Plot filter motifs.}
\usage{
plot_motifs(W_conv, ylow = 0, yhigh = 2, method = "custom",
  plotheight = NULL, fl)
}
\arguments{
\item{W_conv}{Either a list of motifs as ICMs, or a single ICM to visualize as sequence logos.}

\item{ylow}{Minimum Y-axis limit.}

\item{yhigh}{Maximum Y-axis limit.}

\item{method}{Y-label calculation type for sequence logo plot. Either \code{'custom'} or \code{'bits'}.}

\item{plotheight}{Height in inches for output plot.}

\item{fl}{Filename to save plots to. If not provided, a plot object will be returned.}
}
\value{
A ggplot object.
}
\description{
Plot first-layer convolutional filter motifs.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{filter}
\keyword{motif}
\keyword{plot}
