% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizer_defintions.R
\name{position_regularizer}
\alias{position_regularizer}
\title{Position-wise lasso penalty applied to convolutional filters weights.}
\usage{
position_regularizer(weight_matrix)
}
\arguments{
\item{weight_matrix}{Filter weights.}
}
\value{
Lasso penalty.
}
\description{
L1 regularization penalty to constrain individual filter weights with decaying strength nearing the center of tte filter.
}
\seealso{
\code{\link{filter_regularizer}} \code{\link{total_regularizer}} \code{\link{Z_regularizer}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{position}
\keyword{regularizer}
\keyword{sparse}
