% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_definitions.R
\name{restrict_to_ICM}
\alias{restrict_to_ICM}
\title{Constrains weights to be valid information measures.}
\usage{
restrict_to_ICM(wts, offset)
}
\arguments{
\item{wts}{Convolutional filter weights.}

\item{offset}{Filter offset weights.}
}
\value{
Rescaled weights.
}
\description{
Constrains filter weights to be valid ICMs by rescaling the weights.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{ICM}
\keyword{constraint}
\keyword{information}
