% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_definitions.R
\name{restrict_to_IGM}
\alias{restrict_to_IGM}
\title{Constrains weights to be valid information measures.}
\usage{
restrict_to_IGM(wts, offset, bg = c(0.25, 0.25, 0.25, 0.25),
  pseudocount = 0.05)
}
\arguments{
\item{wts}{Convolutional filter weights.}

\item{offset}{Filter offset weights.}

\item{bg}{Vector of background nucleotide probabilities (A, C, G, T)}

\item{pseudocount}{Value to add to zero-valued position probability matrix column sums prior to converting to IGM.}
}
\value{
Rescaled weights.
}
\description{
Constrains filter weights to be valid IGMs by rescaling the weights.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{IGM}
\keyword{constraint}
\keyword{information}
