% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_definitions.R
\name{restrict_to_PWM}
\alias{restrict_to_PWM}
\title{Constrains weights to be valid position weight measures.}
\usage{
restrict_to_PWM(wts, bg = c(0.25, 0.25, 0.25, 0.25),
  pseudoprob = 0.001)
}
\arguments{
\item{wts}{Convolutional filter weights.}

\item{bg}{Vector of background nucleotide probabilities (A, C, G, T)}

\item{pseudoprob}{Value to add to zero-valued position probability matrix entries prior to converting to PWM.}
}
\value{
Rescaled weights.
}
\description{
Constrains filter weights to be valid PWMs by rescaling the weights.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{PWM}
\keyword{constraint}
\keyword{position}
\keyword{weight}
