% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tomtom.R
\name{run_tomtom}
\alias{run_tomtom}
\title{Run Tomtom to match motifs.}
\usage{
run_tomtom(W_conv_list, pfm = F)
}
\arguments{
\item{W_conv_list}{List of motifs to run through Tomtom.}

\item{pfm}{Logical indicating whether position frequency matrices are provided (\code{TRUE}) or information content matrices (\code{FALSE}).}
}
\value{
\code{tomtom_out} directory containing Tomtom output.
}
\description{
Checks estimated filter motifs against JASPAR database motifs and returns significant motif hits. Must have the Memesuit installed \url{http://meme-suite.org/}.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{jaspar}
\keyword{motif}
\keyword{tomtom}
