% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{setup_log_dir}
\alias{setup_log_dir}
\title{Initialize output directory.}
\usage{
setup_log_dir(opt)
}
\arguments{
\item{opt}{Options list.}
}
\value{
NULL. Output directory will be initialized.
}
\description{
Creates the initial output directory, optionally backing up any files which may be overwritten.
}
\examples{
setup_log_dir(list(log_dir='output_directory',cache_Old=FALSE))

}
\seealso{
\code{\link{make_epoch_dir}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{directory}
\keyword{log}
