% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_topN_motif_hits.R
\docType{data}
\name{topN_max_pooling}
\alias{topN_max_pooling}
\title{Top max-pooled hits R6 class wrapper}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
topN_max_pooling
}
\description{
R6 class container for the top max-pooled motif hits
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{callback}
\keyword{model}
