% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizer_defintions.R
\name{total_regularizer}
\alias{total_regularizer}
\title{Total kernel regularization function wrapper.}
\usage{
total_regularizer(weight_matrix)
}
\arguments{
\item{weight_matrix}{Filter weights.}
}
\value{
Total regularization penalty.
}
\description{
Sum of the filter regularization and the position regularization.
}
\seealso{
\code{\link{filter_regularizer}} \code{\link{position_regularizer}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{filter}
\keyword{position}
\keyword{regularizer}
\keyword{sparse}
\keyword{total}
