% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizer_defintions.R
\name{Z_regularizer}
\alias{Z_regularizer}
\title{Regularization of the activation values around 0.5.}
\usage{
Z_regularizer(weight_matrix)
}
\arguments{
\item{weight_matrix}{Filter weights.}
}
\value{
Regularization penalty.
}
\description{
Regularizes the max-pooled activation value (post-sigmoid) around 0.5.
}
\seealso{
\code{\link{filter_regularizer}} \code{\link{position_regularizer}} \code{\link{total_regularizer}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{activation}
\keyword{regularizer}
\keyword{sparse}
