% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_motifs.R
\name{format_filter_motifs}
\alias{format_filter_motifs}
\title{Format filters for motif plotting.}
\usage{
format_filter_motifs(conv_filters)
}
\arguments{
\item{conv_filters}{4-dimensional array containing filter weights.}
}
\value{
List with named motifs.
}
\description{
Convert the information content motifs to a list with named rows for plotting with \code{\link{plot_motifs}}.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{filter}
\keyword{format}
\keyword{motif}
