% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_fixedMotif.R
\name{layer_fixedMotif}
\alias{layer_fixedMotif}
\title{Custom 2-D convolution wrapper for fixed IGMs or PWMs}
\usage{
layer_fixedMotif(object, filters, motif_maxlen, fixed.motifs,
  fixed.offsets, input_shape = NULL, strides = c(1L, 1L),
  padding = "valid", activation = "sigmoid", use_bias = TRUE,
  name = "fixedMotif_conv", trainable = FALSE)
}
\arguments{
\item{object}{Keras model object.}

\item{filters}{Number of convolutional filters.}

\item{motif_maxlen}{Maximum length of the motifs.}

\item{fixed.motifs}{List of motifs to be used as filters.}

\item{fixed.offsets}{Vector of offset values.}

\item{input_shape}{Dimensionality of the input (integer) not including the samples axis. This argument is required when using this layer as the first layer in a model.}

\item{strides}{An integer or list of 2 integers, specifying the strides of the convolution along the width and height. Can be a single integer to specify the same value for all spatial dimensions. Specifying any stride value != 1 is incompatible with specifying any \code{dilation_rate} value != 1.}

\item{padding}{one of "valid" (default) or "same" (case-insensitive).}

\item{activation}{Activation function to use. Default is the sigmoid activation.}

\item{use_bias}{Boolean, whether the layer uses a bias vector. Default is to include a bias.}

\item{name}{An optional name string for the layer. Should be unique in a model (do not reuse the same name twice). It will be autogenerated if it isn't provided. Default is \code{'fixedMotif_conv'}.}

\item{trainable}{Whether the layer weights will be updated during training. Default is \code{FALSE}.}
}
\value{
A Keras model with the added layer.
}
\description{
Custom 2-D convolution wrapper for fixed motifs, such as those coming from a database.
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{conv}
\keyword{convolution}
\keyword{fixedMotif}
\keyword{layer}
